+++ Sensors_driver/helper.h
@@ -4,14 +4,15 @@
 #include <stdint.h>
 #include <stdbool.h>
 
+
 #define RB_MAX_SIZE 100
 
 typedef struct
 {
     uint8_t head;
-    int16_t buffer[RB_MAX_SIZE];
+    float buffer[RB_MAX_SIZE];
 } rb_typedef_t;
 
-void rb_push(rb_typedef_t *rb, int16_t val);
+void rb_push(rb_typedef_t *rb, int32_t val);
 void rb_init(rb_typedef_t *rb);
 #endif
